FS-3235: Move download file into helper module
